requirements.txt
paramiko>=1.16.0,!=2.0.1
PyYAML>=3.1.0
six>=1.9.0
